google_vision/requirements.txt
Flask
python-dotenv
requests
scikit-learn
google-cloud-vision
